GPU (Linux / FreeBSD): unifies DRM related code and code clean up
