+++ src/detection/gpu/gpu_bsd.c
@@ -11,17 +11,6 @@
     #include <bus/pci/pcireg.h> // DragonFly
 #endif
 
-#include "common/library.h"
-#include "util/stringUtils.h"
-#include <xf86drm.h>
-#include <i915_drm.h>
-
-#ifdef FF_HAVE_DRM_AMDGPU
-#include <amdgpu.h>
-#include <amdgpu_drm.h>
-#include <fcntl.h>
-#endif
-
 static void fillGPUTypeGeneric(FFGPUResult* gpu)
 {
     if (gpu->type == FF_GPU_TYPE_UNKNOWN)
@@ -47,137 +36,10 @@ static void fillGPUTypeGeneric(FFGPUResult* gpu)
 }
 
 #if FF_HAVE_DRM
-static const char* drmDetectAmdSpecific(const FFGPUOptions* options, FFGPUResult* gpu, const char* renderPath)
-{
-#if FF_HAVE_DRM_AMDGPU
-    FF_LIBRARY_LOAD(libdrm, "dlopen libdrm_amdgpu" FF_LIBRARY_EXTENSION " failed", "libdrm_amdgpu" FF_LIBRARY_EXTENSION, 1)
-    FF_LIBRARY_LOAD_SYMBOL_MESSAGE(libdrm, amdgpu_device_initialize)
-    FF_LIBRARY_LOAD_SYMBOL_MESSAGE(libdrm, amdgpu_get_marketing_name)
-    FF_LIBRARY_LOAD_SYMBOL_MESSAGE(libdrm, amdgpu_query_gpu_info)
-    FF_LIBRARY_LOAD_SYMBOL_MESSAGE(libdrm, amdgpu_query_sensor_info)
-    FF_LIBRARY_LOAD_SYMBOL_MESSAGE(libdrm, amdgpu_query_heap_info)
-    FF_LIBRARY_LOAD_SYMBOL_MESSAGE(libdrm, amdgpu_device_deinitialize)
+#include "common/library.h"
+#include "util/stringUtils.h"
 
-    FF_AUTO_CLOSE_FD int fd = open(renderPath, O_RDONLY);
-    if (fd < 0) return "Failed to open DRM device";
-
-    amdgpu_device_handle handle;
-    uint32_t majorVersion, minorVersion;
-    if (ffamdgpu_device_initialize(fd, &majorVersion, &minorVersion, &handle) < 0)
-        return "Failed to initialize AMDGPU device";
-
-    ffStrbufAppendF(&gpu->driver, " %u.%u", (unsigned) majorVersion, (unsigned) minorVersion);
-
-    uint32_t value;
-
-    if (options->temp)
-    {
-        if (ffamdgpu_query_sensor_info(handle, AMDGPU_INFO_SENSOR_GPU_TEMP, sizeof(value), &value) >= 0)
-            gpu->temperature = value / 1000.;
-    }
-
-    ffStrbufSetS(&gpu->name, ffamdgpu_get_marketing_name(handle));
-
-    struct amdgpu_gpu_info gpuInfo;
-    if (ffamdgpu_query_gpu_info(handle, &gpuInfo) >= 0)
-    {
-        gpu->coreCount = (int32_t) gpuInfo.cu_active_number;
-        gpu->frequency = (uint32_t) (gpuInfo.max_engine_clk / 1000u);
-        gpu->index = FF_GPU_INDEX_UNSET;
-        gpu->type = gpuInfo.ids_flags & AMDGPU_IDS_FLAGS_FUSION ? FF_GPU_TYPE_INTEGRATED : FF_GPU_TYPE_DISCRETE;
-#define FF_VRAM_CASE(name, value) case value /* AMDGPU_VRAM_TYPE_ ## name */: ffStrbufSetStatic(&gpu->memoryType, #name); break
-        switch (gpuInfo.vram_type)
-        {
-            FF_VRAM_CASE(UNKNOWN, 0);
-            FF_VRAM_CASE(GDDR1, 1);
-            FF_VRAM_CASE(DDR2, 2);
-            FF_VRAM_CASE(GDDR3, 3);
-            FF_VRAM_CASE(GDDR4, 4);
-            FF_VRAM_CASE(GDDR5, 5);
-            FF_VRAM_CASE(HBM, 6);
-            FF_VRAM_CASE(DDR3, 7);
-            FF_VRAM_CASE(DDR4, 8);
-            FF_VRAM_CASE(GDDR6, 9);
-            FF_VRAM_CASE(DDR5, 10);
-            FF_VRAM_CASE(LPDDR4, 11);
-            FF_VRAM_CASE(LPDDR5, 12);
-        default:
-            ffStrbufAppendF(&gpu->memoryType, "Unknown (%u)", gpuInfo.vram_type);
-            break;
-        }
-
-        struct amdgpu_heap_info heapInfo;
-        if (ffamdgpu_query_heap_info(handle, AMDGPU_GEM_DOMAIN_VRAM, 0, &heapInfo) >= 0)
-        {
-            if (gpu->type == FF_GPU_TYPE_DISCRETE)
-            {
-                gpu->dedicated.total = heapInfo.heap_size;
-                gpu->dedicated.used = heapInfo.heap_usage;
-            }
-            else
-            {
-                gpu->shared.total = heapInfo.heap_size;
-                gpu->shared.used = heapInfo.heap_usage;
-            }
-        }
-    }
-
-    if (ffamdgpu_query_sensor_info(handle, AMDGPU_INFO_SENSOR_GPU_LOAD, sizeof(value), &value) >= 0)
-        gpu->coreUsage = value;
-
-    ffamdgpu_device_deinitialize(handle);
-
-    return NULL;
-#else
-    FF_UNUSED(options, gpu, drmKey, buffer);
-    return "Fastfetch is compiled without libdrm support";
-#endif
-}
-
-static const char* drmDetectIntelSpecific(FFGPUResult* gpu, int fd)
-{
-    {
-        int value;
-        drm_i915_getparam_t getparam = { .param = I915_PARAM_EU_TOTAL, .value = &value };
-        if (ioctl(fd, DRM_IOCTL_I915_GETPARAM, &getparam) >= 0)
-            gpu->coreCount = value;
-    }
-    {
-        struct drm_i915_query_item queryItem = {
-            .query_id = DRM_I915_QUERY_MEMORY_REGIONS,
-        };
-        struct drm_i915_query query = {
-            .items_ptr = (uintptr_t) &queryItem,
-            .num_items = 1,
-        };
-        if (ioctl(fd, DRM_IOCTL_I915_QUERY, &query) >= 0 )
-        {
-            FF_AUTO_FREE uint8_t* buffer = calloc(1, (size_t) queryItem.length);
-            queryItem.data_ptr = (uintptr_t) buffer;
-            if (ioctl(fd, DRM_IOCTL_I915_QUERY, &query) >= 0)
-            {
-                gpu->dedicated.total = gpu->shared.total = gpu->dedicated.used = gpu->shared.used = 0;
-                struct drm_i915_query_memory_regions* regionInfo = (void*) buffer;
-                for (uint32_t i = 0; i < regionInfo->num_regions; i++)
-                {
-                    struct drm_i915_memory_region_info* region = regionInfo->regions + i;
-                    switch (region->region.memory_class)
-                    {
-                    case I915_MEMORY_CLASS_SYSTEM:
-                        gpu->shared.total += region->probed_size;
-                        gpu->shared.used += region->probed_size - region->unallocated_size;
-                        break;
-                    case I915_MEMORY_CLASS_DEVICE:
-                        gpu->dedicated.total += region->probed_size;
-                        gpu->dedicated.used += region->probed_size - region->unallocated_size;
-                        break;
-                    }
-                }
-            }
-        }
-    }
-    return NULL;
-}
+#include <xf86drm.h>
 
 static const char* detectByDrm(const FFGPUOptions* options, FFlist* gpus)
 {
@@ -235,43 +97,30 @@ static const char* detectByDrm(const FFGPUOptions* options, FFlist* gpus)
         if (fd < 0) continue;
 
         char driverName[64];
-        char driverDesc[64];
+        driverName[0] = '\0';
         struct drm_version ver = {
             .name = driverName,
             .name_len = ARRAY_SIZE(driverName),
-            .desc = driverDesc,
-            .desc_len = ARRAY_SIZE(driverDesc),
         };
         if (ioctl(fd, DRM_IOCTL_VERSION, &ver) == 0)
             ffStrbufSetF(&gpu->driver, "%*s %d.%d.%d", (int) ver.name_len, ver.name, ver.version_major, ver.version_minor, ver.version_patchlevel);
 
-        if (dev->bustype != DRM_BUS_PCI)
-            continue;
-
-        if (ffStrEquals(ver.name, "i915"))
-            drmDetectIntelSpecific(gpu, fd);
-        else if (ffStrEquals(ver.name, "amdgpu"))
-            drmDetectAmdSpecific(options, gpu, dev->nodes[DRM_NODE_RENDER]);
-        else if (ffStrEquals(ver.name, "nvidia-drm") && (options->temp || options->driverSpecific))
+        if (ffStrStartsWith(driverName, "i915"))
+            ffDrmDetectI915(gpu, fd);
+        else if (ffStrStartsWith(driverName, "amdgpu"))
+            ffDrmDetectAmdgpu(options, gpu, dev->nodes[DRM_NODE_RENDER]);
+        else if (ffStrStartsWith(driverName, "xe"))
+            ffDrmDetectXe(gpu, fd);
+        else if (ffStrStartsWith(driverName, "asahi"))
+            ffDrmDetectAsahi(gpu, fd);
+        else if (dev->bustype == DRM_BUS_PCI)
         {
-            ffDetectNvidiaGpuInfo(&(FFGpuDriverCondition) {
-                                      .type = FF_GPU_DRIVER_CONDITION_TYPE_BUS_ID,
-                                      .pciBusId = {
-                                          .domain = (uint32_t) dev->businfo.pci->domain,
-                                          .bus = dev->businfo.pci->bus,
-                                          .device = dev->businfo.pci->dev,
-                                          .func = dev->businfo.pci->func,
-                                      },
-                                  }, (FFGpuDriverResult) {
-                                      .index = &gpu->index,
-                                      .temp = options->temp ? &gpu->temperature : NULL,
-                                      .memory = options->driverSpecific ? &gpu->dedicated : NULL,
-                                      .coreCount = options->driverSpecific ? (uint32_t*) &gpu->coreCount : NULL,
-                                      .type = &gpu->type,
-                                      .frequency = &gpu->frequency,
-                                      .coreUsage = &gpu->coreUsage,
-                                      .name = &gpu->name,
-                                  }, "libnvidia-ml.so");
+            ffGPUDetectDriverSpecific(options, gpu, (FFGpuDriverPciBusId) {
+                .domain = (uint32_t) dev->businfo.pci->domain,
+                .bus = dev->businfo.pci->bus,
+                .device = dev->businfo.pci->dev,
+                .func = dev->businfo.pci->func,
+            });
         }
 
         if (gpu->name.length == 0)
@@ -333,27 +182,12 @@ static const char* detectByPci(const FFGPUOptions* options, FFlist* gpus)
         gpu->deviceId = (pc->pc_sel.pc_domain * 100000ull) + (pc->pc_sel.pc_bus * 1000ull) + (pc->pc_sel.pc_dev * 10ull) + pc->pc_sel.pc_func;
         gpu->frequency = FF_GPU_FREQUENCY_UNSET;
 
-        if (gpu->vendor.chars == FF_GPU_VENDOR_NAME_NVIDIA && (options->temp || options->driverSpecific))
-        {
-            ffDetectNvidiaGpuInfo(&(FFGpuDriverCondition) {
-                                      .type = FF_GPU_DRIVER_CONDITION_TYPE_BUS_ID,
-                                      .pciBusId = {
-                                          .domain = (uint32_t) pc->pc_sel.pc_domain,
-                                          .bus = pc->pc_sel.pc_bus,
-                                          .device = pc->pc_sel.pc_dev,
-                                          .func = pc->pc_sel.pc_func,
-                                      },
-                                  }, (FFGpuDriverResult) {
-                                      .index = &gpu->index,
-                                      .temp = options->temp ? &gpu->temperature : NULL,
-                                      .memory = options->driverSpecific ? &gpu->dedicated : NULL,
-                                      .coreCount = options->driverSpecific ? (uint32_t*) &gpu->coreCount : NULL,
-                                      .type = &gpu->type,
-                                      .frequency = &gpu->frequency,
-                                      .coreUsage = &gpu->coreUsage,
-                                      .name = &gpu->name,
-                                  }, "libnvidia-ml.so");
-        }
+        ffGPUDetectDriverSpecific(options, gpu, (FFGpuDriverPciBusId) {
+            .domain = (uint32_t) pc->pc_sel.pc_domain,
+            .bus = pc->pc_sel.pc_bus,
+            .device = pc->pc_sel.pc_dev,
+            .func = pc->pc_sel.pc_func,
+        });
 
         if (gpu->name.length == 0)
         {
